+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Change Log
 ==========
 ### 3.4.0
 - clearCache now supports type inference
+- you can now pass `publicPath: true` and it will use the `name` property of your app's `package.json` as the public path
 
 ### 3.3.3
 - array dependencies were being incorrectly flattened